+++ test/report/template_test.dart
@@ -28,8 +28,28 @@ void main() {
       expect(section.summary, contains('No `CHANGELOG.md` found'));
     });
 
+    test('finds missing version in CHANGELOG', () async {
+      final descriptor = package('my_package', version: '2.0.0', extraFiles: [
+        d.file('CHANGELOG.md', '''
+## 1.0.0 First release!
+'''),
+      ]);
+      await descriptor.create();
+      final report = await createReport(PackageContext(
+        sharedContext: SharedAnalysisContext(
+          toolEnvironment: await ToolEnvironment.create(),
+        ),
+        packageDir: descriptor.io.path,
+      ));
+      final section = report.sections
+          .firstWhere((s) => s.title == 'Follow Dart file conventions');
+      expect(section.grantedPoints, 0);
+      expect(section.summary,
+          contains('does not contain reference to the current version'));
+    });
+
     test('Detects insecure links', () async {
-      final descriptor = package('my_package', extraFiles: [
+      final descriptor = package('my_package', version: '1.0.0', extraFiles: [
         d.file('README.md', '''
 # my_package
 Check it out on [github](http://github.com/example/my_package).