+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Do not list SDK packages in the outdated packages report.
 - Fix issue with analyzing packages nested inside git repositories on windows.
   https://github.com/dart-lang/pana/issues/1207 
+- Check the reference of the current version in `CHANGELOG.md`.
 
 ## 0.22.21